Move actor commands to their own directory
