+++ src/hal/actor/commands/calibrations.py
@@ -16,7 +16,7 @@ from . import hal_command_parser, stages
 if TYPE_CHECKING:
     from hal.macros import Macro
 
-    from . import HALCommandType
+    from .. import HALCommandType
 
 
 @hal_command_parser.group()
