+++ src/hal/actor/commands/status.py
